requirements.txt
beir
openai
google-genai
nltk
spacy
rouge_score
timm==0.9.2
loguru
huggingface_hub[cli]